// solar-frontend/src/app/+drivers/models/driver.ts
/**
 * This is Driver interface to give data type for drivers object
 */

export interface IDriver {

  empID: string ;
  bDate: string;
  firstName: string;
  lastName: string;
  middleName: string;
  terminal: string;
  email: string;
  phoneNumber: string;
  passWord: string;
  driverBirthDate:any;
  selectedDate:any;

}